# cmd/docs/openapiv3.yaml
extensionprops: {}
openapi: 3.0.0
components:
  extensionprops: {}
  schemas:
    Device:
      ref: ""
      value:
        extensionprops: {}
        type: object
        properties:
          alias:
            ref: ""
            value:
              extensionprops: {}
              type: string
          common_name:
            ref: ""
            value:
              extensionprops: {}
              type: string
          country:
            ref: ""
            value:
              extensionprops: {}
              type: string
          creation_timestamp:
            ref: ""
            value:
              extensionprops: {}
              type: string
          current_cert_serial_number:
            ref: ""
            value:
              extensionprops: {}
              type: string
          dms_id:
            ref: ""
            value:
              extensionprops: {}
              type: integer
          id:
            ref: ""
            value:
              extensionprops: {}
              type: string
          key_bits:
            ref: ""
            value:
              extensionprops: {}
              type: integer
          key_strength:
            ref: ""
            value:
              extensionprops: {}
              type: string
          key_type:
            ref: ""
            value:
              extensionprops: {}
              type: string
          locality:
            ref: ""
            value:
              extensionprops: {}
              type: string
          organization:
            ref: ""
            value:
              extensionprops: {}
              type: string
          organization_unit:
            ref: ""
            value:
              extensionprops: {}
              type: string
          state:
            ref: ""
            value:
              extensionprops: {}
              type: string
          status:
            ref: ""
            value:
              extensionprops: {}
              type: string
    DeviceCert:
      ref: ""
      value:
        extensionprops: {}
        type: object
        properties:
          common_name:
            ref: ""
            value:
              extensionprops: {}
              type: string
          country:
            ref: ""
            value:
              extensionprops: {}
              type: string
          crt:
            ref: ""
            value:
              extensionprops: {}
              type: string
          device_id:
            ref: ""
            value:
              extensionprops: {}
              type: string
          issuer_name:
            ref: ""
            value:
              extensionprops: {}
              type: string
          locality:
            ref: ""
            value:
              extensionprops: {}
              type: string
          organization:
            ref: ""
            value:
              extensionprops: {}
              type: string
          organization_unit:
            ref: ""
            value:
              extensionprops: {}
              type: string
          serial_number:
            ref: ""
            value:
              extensionprops: {}
              type: string
          state:
            ref: ""
            value:
              extensionprops: {}
              type: string
          status:
            ref: ""
            value:
              extensionprops: {}
              type: string
          valid_from:
            ref: ""
            value:
              extensionprops: {}
              type: string
          valid_to:
            ref: ""
            value:
              extensionprops: {}
              type: string
    DeviceCertHistory:
      ref: ""
      value:
        extensionprops: {}
        type: object
        properties:
          creation_timestamp:
            ref: ""
            value:
              extensionprops: {}
              type: string
          device_id:
            ref: ""
            value:
              extensionprops: {}
              type: string
          issuer_name:
            ref: ""
            value:
              extensionprops: {}
              type: string
          issuer_serial_number:
            ref: ""
            value:
              extensionprops: {}
              type: string
          serial_number:
            ref: ""
            value:
              extensionprops: {}
              type: string
          status:
            ref: ""
            value:
              extensionprops: {}
              type: string
    DeviceLog:
      ref: ""
      value:
        extensionprops: {}
        type: object
        properties:
          device_id:
            ref: ""
            value:
              extensionprops: {}
              type: string
          id:
            ref: ""
            value:
              extensionprops: {}
              type: string
          log_message:
            ref: ""
            value:
              extensionprops: {}
              type: string
          log_type:
            ref: ""
            value:
              extensionprops: {}
              type: string
          timestamp:
            ref: ""
            value:
              extensionprops: {}
              type: string
  requestBodies:
    PostDeviceRequest:
      ref: ""
      value:
        extensionprops: {}
        description: Request used for creating a new Certificate Authority
        required: true
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                properties:
                  subject:
                    ref: '#/components/schemas/Device'
                    value: null
  responses:
    DeleteRevokeResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after revoking a device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
    DeviceResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after creating a device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: '#/components/schemas/Device'
              value: null
    ErrorResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response when errors happen.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                properties:
                  error:
                    ref: ""
                    value:
                      extensionprops: {}
                      type: string
    GetDeviceCertHistoryResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after getting certificate history of a
          device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: '#/components/schemas/DeviceCertHistory'
              value: null
    GetDeviceCertResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after getting certificate of a device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: '#/components/schemas/DeviceCert'
              value: null
    GetDeviceLogsResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after getting logs of a device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                type: array
                items:
                  ref: '#/components/schemas/DeviceLog'
                  value: null
    GetDeviceResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after creating a device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                type: array
                items:
                  ref: '#/components/schemas/Device'
                  value: null
    GetDmsLastIssueCertResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after getting last iisued certificate
          of a device.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: '#/components/schemas/DeviceCertHistory'
              value: null
    HealthResponse:
      ref: ""
      value:
        extensionprops: {}
        description: Response returned back after healthchecking.
        content:
          application/json:
            extensionprops: {}
            schema:
              ref: ""
              value:
                extensionprops: {}
                properties:
                  healthy:
                    ref: ""
                    value:
                      extensionprops: {}
                      type: boolean
info:
  extensionprops: {}
  title: Lamassu Device Manager API
  description: REST API used for interacting with Lamassu Device Manager
  contact:
    extensionprops: {}
    url: https://github.com/lamassuiot
  license:
    extensionprops: {}
    name: MPL v2.0
    url: https://github.com/lamassuiot/lamassu-compose/blob/main/LICENSE
  version: 0.0.0
paths:
  /v1/devices:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get Devices
      operationId: GetDevices
      responses:
        "200":
          ref: '#/components/responses/GetDeviceResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
    post:
      extensionprops: {}
      description: Post Device
      operationId: PostDevice
      requestBody:
        ref: '#/components/requestBodies/PostDeviceRequest'
        value: null
      responses:
        "200":
          ref: '#/components/responses/DeviceResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/{deviceId}:
    extensionprops: {}
    delete:
      extensionprops: {}
      description: Delete Device By Id
      operationId: DeleteDevice
      parameters:
      - ref: ""
        value:
          extensionprops: {}
          name: deviceId
          in: path
          required: true
          schema:
            ref: ""
            value:
              extensionprops: {}
              type: string
      responses:
        "200":
          ref: '#/components/responses/StringResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
    get:
      extensionprops: {}
      description: Get Device By Id
      operationId: GetDeviceById
      parameters:
      - ref: ""
        value:
          extensionprops: {}
          name: deviceId
          in: path
          required: true
          schema:
            ref: ""
            value:
              extensionprops: {}
              type: string
      responses:
        "200":
          ref: '#/components/responses/DeviceResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/{deviceId}/cert:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get Device Cert of deviceId
      operationId: GetDeviceCert
      parameters:
      - ref: ""
        value:
          extensionprops: {}
          name: deviceId
          in: path
          required: true
          schema:
            ref: ""
            value:
              extensionprops: {}
              type: string
      responses:
        "200":
          ref: '#/components/responses/GetDeviceCertResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/{deviceId}/cert-history:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get Device Cert History of deviceId
      operationId: GetDeviceCertHistory
      parameters:
      - ref: ""
        value:
          extensionprops: {}
          name: deviceId
          in: path
          required: true
          schema:
            ref: ""
            value:
              extensionprops: {}
              type: string
      responses:
        "200":
          ref: '#/components/responses/GetDeviceCertHistoryResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/{deviceId}/logs:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get Device Logs of deviceId
      operationId: GetDeviceLogs
      parameters:
      - ref: ""
        value:
          extensionprops: {}
          name: deviceId
          in: path
          required: true
          schema:
            ref: ""
            value:
              extensionprops: {}
              type: string
      responses:
        "200":
          ref: '#/components/responses/GetDeviceLogsResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/{deviceId}/revoke:
    extensionprops: {}
    delete:
      extensionprops: {}
      description: Delete Revoke device by Id
      operationId: DeleteRevoke
      parameters:
      - ref: ""
        value:
          extensionprops: {}
          name: deviceId
          in: path
          required: true
          schema:
            ref: ""
            value:
              extensionprops: {}
              type: string
      responses:
        "200":
          ref: '#/components/responses/DeleteRevokeResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/dms-cert-history/last-issued:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get Dms Cert History of last Thirty Days
      operationId: GetDmsLastIssueCert
      responses:
        "200":
          ref: '#/components/responses/GetDmsLastIssueCertResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/devices/dms-cert-history/thirty-days:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get Dms Cert History of last Thirty Days
      operationId: GetDmsCertHistoryThirtyDays
      responses:
        "200":
          ref: '#/components/responses/GetDeviceCertHistoryResponse'
          value: null
        "400":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "401":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "403":
          ref: '#/components/responses/ErrorResponse'
          value: null
        "500":
          ref: '#/components/responses/ErrorResponse'
          value: null
  /v1/health:
    extensionprops: {}
    get:
      extensionprops: {}
      description: Get health status
      operationId: Health
      responses:
        "200":
          ref: '#/components/responses/HealthResponse'
          value: null
servers:
- extensionprops: {}
  url: /
  description: Current Server